refactoring and adding featured logic into a reusable hook
