+++ apps/website/components/templates/landing/featured/index.ts
@@ -1,0 +1,1 @@
+export * from './featured';
