--- apps/website/components/templates/landing/for-users/index.ts
@@ -1,1 +1,0 @@
-export * from './for-users';
